Refactored the search function in the list component
